+++ src/main/java/io/kyberorg/yalsee/json/QRCodeResponse.java
@@ -10,7 +10,7 @@ import lombok.Data;
  * @since 2.6
  */
 @Data(staticConstructor = "withQRCode")
-public class QRCodeResponseJson implements YalseeJson {
+public class QRCodeResponse implements YalseeJson {
     @JsonProperty("qr_code")
     private final String qrCode;